backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
python-multipart
pymupdf
python-docx
sentence-transformers
chromadb
google-generativeai
python-dotenv
pydantic[email]
